polling::epoll

Enum Notifier

enum Notifier {
    EventFd(OwnedFd),
    Pipe {
        read_pipe: OwnedFd,
        write_pipe: OwnedFd,
    },
}

The notifier for Linux.

Certain container runtimes do not expose eventfd to the client, as it relies on the host and can be used to “escape” the container under certain conditions. Gramine is the prime example, see here. In this case, fall back to using a pipe.

Variants

EventFd(OwnedFd)

The primary notifier, using eventfd.

Pipe

The fallback notifier, using a pipe.

Fields

read_pipe: OwnedFd

The read end of the pipe.

write_pipe: OwnedFd

The write end of the pipe.

Implementations

impl Notifier

fn new() -> Result<Self>

Create a new notifier.

fn as_fd(&self) -> BorrowedFd<'_>

The file descriptor to register in the poller.

fn notify(&self)

Notify the poller.

fn clear(&self)

Clear the notification.