polling::epoll

Struct Poller

pub struct Poller {
    epoll_fd: OwnedFd,
    notifier: Notifier,
    timer_fd: Option<OwnedFd>,
}

Interface to epoll.

Fields
epoll_fd: OwnedFd

File descriptor for the epoll instance.

notifier: Notifier

Notifier used to wake up epoll.

timer_fd: Option<OwnedFd>

File descriptor for the timerfd that produces timeouts.

Redox does not support timerfd.

Implementations

impl Poller

pub fn new() -> Result<Poller>

Creates a new poller.

pub fn supports_level(&self) -> bool

Whether this poller supports level-triggered events.

pub fn supports_edge(&self) -> bool

Whether the poller supports edge-triggered events.

pub unsafe fn add(&self, fd: RawFd, ev: Event, mode: PollMode) -> Result<()>

Adds a new file descriptor.

Safety

The fd must be a valid file descriptor. The usual condition of remaining registered in the Poller doesn’t apply to epoll.

pub fn modify( &self, fd: BorrowedFd<'_>, ev: Event, mode: PollMode, ) -> Result<()>

Modifies an existing file descriptor.

pub fn delete(&self, fd: BorrowedFd<'_>) -> Result<()>

Deletes a file descriptor.

pub fn wait_deadline( &self, events: &mut Events, deadline: Option<Instant>, ) -> Result<()>

Waits for I/O events with an optional deadline.

pub fn notify(&self) -> Result<()>

Sends a notification to wake up the current or next wait() call.